Rename: EditProfile 컴포넌트 분리 #63
